2018 minimal
