+++ assets/jarallax/jarallax.js
@@ -4,454 +4,16 @@
  * Author  : _nK http://nkdev.info
  * GitHub  : https://github.com/nk-o/jarallax
  */
-(function(factory) {
-    'use strict';
-    if (typeof define === 'function' && define.amd) {
-        define(['jquery'], factory);
-    } else if (typeof exports !== 'undefined') {
-        module.exports = factory(require('jquery'));
-    } else {
-        factory(jQuery);
-    }
-}(function($) {
-    // Adapted from https://gist.github.com/paulirish/1579671
-    if (!Date.now)
-        Date.now = function() { return new Date().getTime(); };
-    if(!window.requestAnimationFrame)
-        (function() {
-            'use strict';
-            
-            var vendors = ['webkit', 'moz'];
-            for (var i = 0; i < vendors.length && !window.requestAnimationFrame; ++i) {
-                var vp = vendors[i];
-                window.requestAnimationFrame = window[vp+'RequestAnimationFrame'];
-                window.cancelAnimationFrame = (window[vp+'CancelAnimationFrame']
-                                           || window[vp+'CancelRequestAnimationFrame']);
-            }
-            if (/iP(ad|hone|od).*OS 6/.test(window.navigator.userAgent) // iOS6 is buggy
-                || !window.requestAnimationFrame || !window.cancelAnimationFrame) {
-                var lastTime = 0;
-                window.requestAnimationFrame = function(callback) {
-                    var now = Date.now();
-                    var nextTime = Math.max(lastTime + 16, now);
-                    return setTimeout(function() { callback(lastTime = nextTime); },
-                                      nextTime - now);
-                };
-                window.cancelAnimationFrame = clearTimeout;
-            }
-        }());
-
-    var supportTransform = (function() {
-        var prefixes = 'transform WebkitTransform MozTransform OTransform msTransform'.split(' ');
-        var div = document.createElement('div');
-        for(var i = 0; i < prefixes.length; i++) {
-            if(div && div.style[prefixes[i]] !== undefined) {
-                return prefixes[i];
-            }
-        }
-        return false;
-    }());
-
-    var support3dtransform = (function() {
-        if (!window.getComputedStyle) {
-            return false;
-        }
-
-        var el = document.createElement('p'), 
-            has3d,
-            transforms = {
-                'webkitTransform':'-webkit-transform',
-                'OTransform':'-o-transform',
-                'msTransform':'-ms-transform',
-                'MozTransform':'-moz-transform',
-                'transform':'transform'
-            };
-
-        // Add it to the body to get the computed style.
-        (document.body || document.documentElement).insertBefore(el, null);
-
-        for (var t in transforms) {
-            if (el.style[t] !== undefined) {
-                el.style[t] = "translate3d(1px,1px,1px)";
-                has3d = window.getComputedStyle(el).getPropertyValue(transforms[t]);
-            }
-        }
-
-        (document.body || document.documentElement).removeChild(el);
-
-        return (has3d !== undefined && has3d.length > 0 && has3d !== "none");
-    }());
-    
-    var isAndroid = navigator.userAgent.toLowerCase().indexOf('android') > -1;
-    var isOperaOld = !!window.opera;
-
-    // list with all jarallax instances
-    // need to render all in one scroll/resize event
-    var jarallaxList = [];
-
-    // Jarallax instance
-    var Jarallax = (function() {
-        var instanceID = 0;
-
-        function Jarallax(item, userOptions) {
-            var _this = this,
-                dataOptions;
-
-            _this.$item      = $(item);
-
-            _this.defaults   = {
-                speed             : 0.5,
-                imgSrc            : null,
-                imgWidth          : null,
-                imgHeight         : null,
-                enableTransform   : true,
-                zIndex            : -100
-            };
-            dataOptions      = _this.$item.data('jarallax') || {};
-            _this.options    = $.extend({}, _this.defaults, dataOptions, userOptions);
-
-            // fix speed option [0.0, 1.0]
-            _this.options.speed = Math.min(1, Math.max(0, parseFloat(_this.options.speed)));
-
-            _this.instanceID = instanceID++;
-
-            _this.image      = {
-                src        : _this.options.imgSrc || null,
-                $container : null,
-                $item      : null,
-                width      : _this.options.imgWidth || null,
-                height     : _this.options.imgHeight || null,
-                // fix for Android devices
-                // use <img> instead background image - more smoothly
-                useImgTag  : isAndroid || isOperaOld
-            }
-
-            if(_this.initImg()) {
-                _this.init();
-
-                jarallaxList.push(_this);
-            }
-        }
-
-        return Jarallax;
-    }());
-
-    Jarallax.prototype.initImg = function() {
-        var _this = this;
-
-        // get image src
-        if(_this.image.src === null) {
-            _this.image.src = _this.$item.css('background-image').replace(/^url\(['"]?/g,'').replace(/['"]?\)$/g,'');
-        }
-        if(!_this.image.src || _this.image.src === 'none') {
-            return false;
-        }
-        return true;
-    }
-
-    Jarallax.prototype.init = function() {
-        var _this = this,
-            containerStyles = {
-                position         : 'absolute',
-                top              : 0,
-                left             : 0,
-                width            : '100%',
-                height           : '100%',
-                overflow         : 'hidden',
-                'pointer-events' : 'none',
-                'transition'     : 'transform linear -1ms, -webkit-transform linear -1ms'
-            },
-            imageStyles = {
-                position              : 'fixed'
-            };
-
-        // container for parallax image
-        _this.image.$container = $('<div>')
-            .css(containerStyles)
-            .css({
-                visibility : 'hidden',
-                'z-index'  : _this.options.zIndex
-            })
-            .attr('id', 'jarallax-container-' + _this.instanceID)
-            .prependTo(_this.$item);
-
-        // use img tag
-        if(_this.image.useImgTag && supportTransform) {
-            _this.image.$item = $('<img>').attr('src', _this.image.src);
-            imageStyles = $.extend({
-                'max-width' : 'none'
-            }, containerStyles, imageStyles)
-        }
-
-        // use div with background image
-        else {
-            _this.image.$item = $('<div>');
-            imageStyles = $.extend({
-                'background-position' : '50% 50%',
-                'background-repeat'   : 'no-repeat no-repeat',
-                'background-image'    : 'url("' + _this.image.src + '")'
-            }, containerStyles, imageStyles)
-        }
-
-        // parallax image
-        _this.image.$item.css(imageStyles)
-            .prependTo(_this.image.$container);
-
-        // cover image if width and height is ready
-        function initAfterReady() {
-            _this.coverImage();
-            _this.clipContainer();
-            _this.onScroll(true);
-
-            // save default user styles
-            _this.$item.data('jarallax-original-styles', _this.$item.attr('style'));
-
-            // timeout to fix IE blinking
-            setTimeout(function() {
-                // remove default user background
-                _this.$item.css({
-                    'background-image'      : 'none',
-                    'background-attachment' : 'scroll',
-                    'background-size'       : 'auto'
-                });
-            }, 0);
-        }
-
-        if(_this.image.width && _this.image.height) {
-            // init if width and height already exists
-            initAfterReady();
-        } else {
-            // load image and get width and height
-            _this.getImageSize(_this.image.src, function(width, height) {
-                _this.image.width  = width;
-                _this.image.height = height;
-                initAfterReady();
-            });
-        }
-    };
-
-    Jarallax.prototype.destroy = function() {
-        var _this = this;
-
-        // remove from instances list
-        for(var k = 0, len = jarallaxList.length; k < len; k++) {
-            if(jarallaxList[k].instanceID === _this.instanceID) {
-                jarallaxList.splice(k, 1);
-                break;
-            }
-        }
-
-        // remove additional styles for clip
-        $('head #jarallax-clip-' + _this.instanceID).remove();
-
-        _this.$item.attr('style', _this.$item.data('jarallax-original-styles'));
-        _this.$item.removeData('jarallax-original-styles');
-
-        _this.image.$container.remove();
-
-        delete _this.$item[0].jarallax;
-    }
-
-    // round to 2 decimals
-    Jarallax.prototype.round = function(num) {
-        return Math.floor(num * 100) / 100;
-    }
-
-    Jarallax.prototype.getImageSize = function(src, callback) {
-        if(!src || !callback) {
-            return false;
-        }
-
-        var tempImg = new Image();
-        tempImg.onload = function() {
-            callback(tempImg.width, tempImg.height)
-        }
-        tempImg.src = src;
-    }
-
-    // it will remove some image overlapping
-    // overlapping occur due to an image position fixed inside absolute possition element (webkit based browsers works without any fix)
-    Jarallax.prototype.clipContainer = function() {
-        var _this  = this,
-            width  = _this.image.$container.outerWidth(true),
-            height = _this.image.$container.outerHeight(true);
-
-        var $styles = $('head #jarallax-clip-' + _this.instanceID);
-        if(!$styles.length) {
-            $('head').append('<style type="text/css" id="jarallax-clip-' + _this.instanceID + '"></style>');
-            $styles = $('head #jarallax-clip-' + _this.instanceID);
-        }
-
-        var css = [
-            '#jarallax-container-' + _this.instanceID + ' {',
-            '   clip: rect(0px ' + width + 'px ' + height + 'px 0);',
-            '   clip: rect(0px, ' + width + 'px, ' + height + 'px, 0);',
-            '}'
-        ].join('\n');
-
-        // add clip styles inline (this method need for support IE8 and less browsers)
-        if ($styles[0].styleSheet) {
-            $styles[0].styleSheet.cssText = css;
-        } else {
-            $styles.html(css);
-        }
-    }
-
-    Jarallax.prototype.coverImage = function() {
-        var _this = this;
-
-        if(!_this.image.width || !_this.image.height) {
-            return;
-        }
-
-        var contW = _this.image.$container.outerWidth(true),
-            contH = _this.image.$container.outerHeight(true),
-            wndW  = $(window).outerWidth(true),
-            whdH  = $(window).outerHeight(true),
-            imgW  = _this.image.width,
-            imgH  = _this.image.height,
-            resultWidth, resultHeight;
-
-        var css = {
-            width  : Math.max(wndW, contW) * Math.max(_this.options.speed, 1),
-            height : Math.max(whdH, contH) * Math.max(_this.options.speed, 1)
-        };
-
-        // cover by width
-        if(css.width / css.height > imgW / imgH) {
-            resultWidth = css.width;
-            resultHeight = css.width * imgH / imgW;
-        }
-
-        // cover by height
-        else {
-            resultWidth = css.height * imgW / imgH;
-            resultHeight = css.height;
-        }
-        
-        // for img tag
-        if(_this.image.useImgTag && supportTransform) {
-            css.width = _this.round(resultWidth);
-            css.height = _this.round(resultHeight);
-            css.marginLeft = _this.round(- (resultWidth - contW) / 2);
-            css.marginTop = _this.round(- (resultHeight - contH) / 2);
-        }
-
-        // for div with background image
-        else {
-            css.backgroundSize = _this.round(resultWidth) + 'px ' + _this.round(resultHeight) + 'px';
-        }
-
-        // apply to item
-        _this.image.$item.css(css);
-    };
-
-    Jarallax.prototype.onScroll = function(force) {
-        var _this = this;
-
-        if(!_this.image.width || !_this.image.height) {
-            return;
-        }
-
-        var scrollTop     = $(window).scrollTop(),
-            scrollLeft    = $(window).scrollLeft(),
-            wndWidth      = $(window).width(),
-            wndHeight     = $(window).height(),
-            // starting position of each element to have parallax applied to it
-            sectionTop    = _this.$item.offset().top,
-            sectionLeft    = _this.$item.offset().left,
-            sectionHeight = _this.$item.outerHeight(true),
-            css           = {
-                visibility         : 'visible',
-                backgroundPosition : '50% 50%'
-            };
-
-        // Check if totally above or totally below viewport
-        var check = force ? false
-                          : sectionTop + sectionHeight < scrollTop || sectionTop > scrollTop + wndHeight;
-        if (check) {
-            return;
-        }
-
-        // calculate parallax
-        var positionY = - (scrollTop - sectionTop) * _this.options.speed;
-        var positionX = - (scrollLeft - sectionLeft) * _this.options.speed;
-            positionY = _this.round(positionY);
-            positionX = _this.round(positionX);
-
-        if(supportTransform && _this.options.enableTransform) {
-            css.transform = 'translateY(' + positionY + 'px) translateX(' + positionX + 'px)';
-            if(support3dtransform) {
-                css.transform = 'translate3d(' + positionX + 'px, ' + positionY + 'px, 0)';
-            }
-        } else {
-            css.backgroundPosition = positionX + 'px ' + positionY + 'px';
-        }
-
-        _this.image.$item.css(css);
-    };
-
-    // init events
-    (function() {
-        $(window).on('scroll.jarallax', function() {
-            window.requestAnimationFrame(function() {
-                for(var k = 0, len = jarallaxList.length; k < len; k++) {
-                    jarallaxList[k].onScroll();
-                }
-            });
-        });
-
-        var timeout;
-        $(window).on('resize.jarallax orientationchange.jarallax load.jarallax', function() {
-            clearTimeout(timeout);
-            timeout = setTimeout(function() {
-                window.requestAnimationFrame(function() {
-                    for(var k = 0, len = jarallaxList.length; k < len; k++) {
-                        var _this = jarallaxList[k];
-                        _this.coverImage();
-                        _this.clipContainer();
-                        _this.onScroll();
-                    }
-                });
-            }, 100);
-        });
-    }());
-
-    var oldJarallax = $.fn.jarallax;
-
-    $.fn.jarallax = function() {
-        var items = this,
-            options = arguments[0],
-            args = Array.prototype.slice.call(arguments, 1),
-            len = items.length,
-            k = 0,
-            ret;
-
-        for (k; k < len; k++) {
-            if (typeof options === 'object' || typeof options === 'undefined') {
-                if(!items[k].jarallax) {
-                    items[k].jarallax = new Jarallax(items[k], options);
-                }
-            }
-            else {
-                ret = items[k].jarallax ? items[k].jarallax[options].apply(items[k].jarallax, args) : undefined;
-            }
-            if (typeof ret !== 'undefined') {
-                return ret;
-            }
-        }
-
-        return this;
-    };
-
-    // no conflict
-    $.fn.jarallax.noConflict = function () {
-        $.fn.jarallax = oldJarallax;
-        return this;
-    };
-
-    // data-jarallax initialization
-    $(document).on('ready.data-jarallax', function () {
-        $('[data-jarallax]').jarallax();
-    });
-}));
+(function(e){"function"===typeof define&&define.amd?define(["jquery"],e):"undefined"!==typeof exports?module.exports=e(require("jquery")):e(jQuery)})(function(e){Date.now||(Date.now=function(){return(new Date).getTime()});window.requestAnimationFrame||function(){for(var b=["webkit","moz"],a=0;a<b.length&&!window.requestAnimationFrame;++a){var c=b[a];window.requestAnimationFrame=window[c+"RequestAnimationFrame"];window.cancelAnimationFrame=window[c+"CancelAnimationFrame"]||window[c+"CancelRequestAnimationFrame"]}if(/iP(ad|hone|od).*OS 6/.test(window.navigator.userAgent)||
+!window.requestAnimationFrame||!window.cancelAnimationFrame){var d=0;window.requestAnimationFrame=function(a){var b=Date.now(),c=Math.max(d+16,b);return setTimeout(function(){a(d=c)},c-b)};window.cancelAnimationFrame=clearTimeout}}();var l=function(){for(var b=["transform","WebkitTransform","MozTransform","OTransform","msTransform"],a=document.createElement("div"),c=0;c<b.length;c++)if(a&&void 0!==a.style[b[c]])return b[c];return!1}(),m=function(){if(!window.getComputedStyle)return!1;var b=document.createElement("p"),
+a,c={webkitTransform:"-webkit-transform",OTransform:"-o-transform",msTransform:"-ms-transform",MozTransform:"-moz-transform",transform:"transform"};(document.body||document.documentElement).insertBefore(b,null);for(var d in c)void 0!==b.style[d]&&(b.style[d]="translate3d(1px,1px,1px)",a=window.getComputedStyle(b).getPropertyValue(c[d]));(document.body||document.documentElement).removeChild(b);return void 0!==a&&0<a.length&&"none"!==a}(),n=-1<navigator.userAgent.toLowerCase().indexOf("android"),p=
+!!window.opera,h=[],f=function(){var b=0;return function(a,c){var d;this.$item=e(a);this.defaults={speed:.5,imgSrc:null,imgWidth:null,imgHeight:null,enableTransform:!0,zIndex:-100};d=this.$item.data("jarallax")||{};this.options=e.extend({},this.defaults,d,c);this.options.speed=Math.min(1,Math.max(0,parseFloat(this.options.speed)));this.instanceID=b++;this.image={src:this.options.imgSrc||null,$container:null,$item:null,width:this.options.imgWidth||null,height:this.options.imgHeight||null,useImgTag:n||
+p};this.initImg()&&(this.init(),h.push(this))}}();f.prototype.initImg=function(){null===this.image.src&&(this.image.src=this.$item.css("background-image").replace(/^url\(['"]?/g,"").replace(/['"]?\)$/g,""));return this.image.src&&"none"!==this.image.src?!0:!1};f.prototype.init=function(){function b(){a.coverImage();a.clipContainer();a.onScroll(!0);a.$item.data("jarallax-original-styles",a.$item.attr("style"));setTimeout(function(){a.$item.css({"background-image":"none","background-attachment":"scroll",
+"background-size":"auto"})},0)}var a=this,c={position:"absolute",top:0,left:0,width:"100%",height:"100%",overflow:"hidden","pointer-events":"none",transition:"transform linear -1ms, -webkit-transform linear -1ms"},d={position:"fixed"};a.image.$container=e("<div>").css(c).css({visibility:"hidden","z-index":a.options.zIndex}).attr("id","jarallax-container-"+a.instanceID).prependTo(a.$item);a.image.useImgTag&&l?(a.image.$item=e("<img>").attr("src",a.image.src),d=e.extend({"max-width":"none"},c,d)):(a.image.$item=
+e("<div>"),d=e.extend({"background-position":"50% 50%","background-repeat":"no-repeat no-repeat","background-image":'url("'+a.image.src+'")'},c,d));a.image.$item.css(d).prependTo(a.image.$container);a.image.width&&a.image.height?b():a.getImageSize(a.image.src,function(c,d){a.image.width=c;a.image.height=d;b()})};f.prototype.destroy=function(){for(var b=0,a=h.length;b<a;b++)if(h[b].instanceID===this.instanceID){h.splice(b,1);break}e("head #jarallax-clip-"+this.instanceID).remove();this.$item.attr("style",
+this.$item.data("jarallax-original-styles"));this.$item.removeData("jarallax-original-styles");this.image.$container.remove();delete this.$item[0].jarallax};f.prototype.round=function(b){return Math.floor(100*b)/100};f.prototype.getImageSize=function(b,a){if(!b||!a)return!1;var c=new Image;c.onload=function(){a(c.width,c.height)};c.src=b};f.prototype.clipContainer=function(){var b=this.image.$container.outerWidth(!0),a=this.image.$container.outerHeight(!0),c=e("head #jarallax-clip-"+this.instanceID);
+c.length||(e("head").append('<style type="text/css" id="jarallax-clip-'+this.instanceID+'"></style>'),c=e("head #jarallax-clip-"+this.instanceID));b=["#jarallax-container-"+this.instanceID+" {","   clip: rect(0px "+b+"px "+a+"px 0);","   clip: rect(0px, "+b+"px, "+a+"px, 0);","}"].join("\n");c[0].styleSheet?c[0].styleSheet.cssText=b:c.html(b)};f.prototype.coverImage=function(){if(this.image.width&&this.image.height){var b=this.image.$container.outerWidth(!0),a=this.image.$container.outerHeight(!0),
+c=e(window).outerWidth(!0),d=e(window).outerHeight(!0),g=this.image.width,f=this.image.height,d={width:Math.max(c,b)*Math.max(this.options.speed,1),height:Math.max(d,a)*Math.max(this.options.speed,1)};d.width/d.height>g/f?(c=d.width,g=d.width*f/g):(c=d.height*g/f,g=d.height);this.image.useImgTag&&l?(d.width=this.round(c),d.height=this.round(g),d.marginLeft=this.round(-(c-b)/2),d.marginTop=this.round(-(g-a)/2)):d.backgroundSize=this.round(c)+"px "+this.round(g)+"px";this.image.$item.css(d)}};f.prototype.onScroll=
+function(b){if(this.image.width&&this.image.height){var a=e(window).scrollTop(),c=e(window).scrollLeft();e(window).width();var d=e(window).height(),g=this.$item.offset().top,f=this.$item.offset().left,h=this.$item.outerHeight(!0),k={visibility:"visible",backgroundPosition:"50% 50%"};(b?0:g+h<a||g>a+d)||(b=-(a-g)*this.options.speed,c=-(c-f)*this.options.speed,b=this.round(b),c=this.round(c),l&&this.options.enableTransform?(k.transform="translateY("+b+"px) translateX("+c+"px)",m&&(k.transform="translate3d("+
+c+"px, "+b+"px, 0)")):k.backgroundPosition=c+"px "+b+"px",this.image.$item.css(k))}};(function(){e(window).on("scroll.jarallax",function(){window.requestAnimationFrame(function(){for(var a=0,b=h.length;a<b;a++)h[a].onScroll()})});var b;e(window).on("resize.jarallax orientationchange.jarallax load.jarallax",function(){clearTimeout(b);b=setTimeout(function(){window.requestAnimationFrame(function(){for(var a=0,b=h.length;a<b;a++){var d=h[a];d.coverImage();d.clipContainer();d.onScroll()}})},100)})})();
+var q=e.fn.jarallax;e.fn.jarallax=function(){var b=arguments[0],a=Array.prototype.slice.call(arguments,1),c=this.length,d=0,e;for(d;d<c;d++)if("object"===typeof b||"undefined"===typeof b?this[d].jarallax||(this[d].jarallax=new f(this[d],b)):e=this[d].jarallax?this[d].jarallax[b].apply(this[d].jarallax,a):void 0,"undefined"!==typeof e)return e;return this};e.fn.jarallax.noConflict=function(){e.fn.jarallax=q;return this};e(document).on("ready.data-jarallax",function(){e("[data-jarallax]").jarallax()})});
