+++ assets/mobirise/js/script.js
@@ -1,421 +1,17 @@
-!function(){try{document.getElementsByClassName("engine")[0].getElementsByTagName("a")[0].removeAttribute("rel")}catch(b){}if(!document.getElementById("top-1")){var a=document.createElement("section");a.id="top-1";a.className="engine";a.innerHTML='<a href="https://mobirise.info">Mobirise</a> Mobirise v4.3.4';document.body.insertBefore(a,document.body.childNodes[0])}}();
-(function($){
 
-    $.extend($.easing, {
-        easeInOutCubic : function(x, t, b, c, d){
-            if ((t/=d/2) < 1) return c/2*t*t*t + b;
-            return c/2*((t-=2)*t*t + 2) + b;
-        }
-    });
-
-    $.fn.outerFind = function(selector){
-        return this.find(selector).addBack(selector);
-    };
-
-    (function($,sr){
-        // debouncing function from John Hann
-        // http://unscriptable.com/index.php/2009/03/20/debouncing-javascript-methods/
-        var debounce = function (func, threshold, execAsap) {
-            var timeout;
-
-            return function debounced () {
-                var obj = this, args = arguments;
-                function delayed () {
-                    if (!execAsap) func.apply(obj, args);
-                    timeout = null;
-                };
-
-                if (timeout) clearTimeout(timeout);
-                else if (execAsap) func.apply(obj, args);
-
-                timeout = setTimeout(delayed, threshold || 100);
-            };
-        }
-        // smartresize 
-        jQuery.fn[sr] = function(fn){  return fn ? this.bind('resize', debounce(fn)) : this.trigger(sr); };
-
-    })(jQuery,'smartresize');
-
-    (function(){
-        
-        var scrollbarWidth = 0, originalMargin, touchHandler = function(event){
-            event.preventDefault();
-        };
-
-        function getScrollbarWidth(){
-            if (scrollbarWidth) return scrollbarWidth;
-            var scrollDiv = document.createElement('div');
-            $.each({
-                top : '-9999px',
-                width  : '50px',
-                height : '50px',
-                overflow : 'scroll', 
-                position : 'absolute'
-            }, function(property, value){
-                scrollDiv.style[property] = value;
-            });
-            $('body').append(scrollDiv);
-            scrollbarWidth = scrollDiv.offsetWidth - scrollDiv.clientWidth;
-            $('body')[0].removeChild(scrollDiv);
-            return scrollbarWidth;
-        }
-
-    })();
-
-    $.isMobile = function(type){
-        var reg = [];
-        var any = {
-            blackberry : 'BlackBerry',
-            android : 'Android',
-            windows : 'IEMobile',
-            opera : 'Opera Mini',
-            ios : 'iPhone|iPad|iPod'
-        };
-        type = 'undefined' == $.type(type) ? '*' : type.toLowerCase();
-        if ('*' == type) reg = $.map(any, function(v){ return v; });
-        else if (type in any) reg.push(any[type]);
-        return !!(reg.length && navigator.userAgent.match(new RegExp(reg.join('|'), 'i')));
-    };
-
-    var isSupportViewportUnits = (function(){
-        // modernizr implementation
-        var $elem = $('<div style="height: 50vh; position: absolute; top: -1000px; left: -1000px;">').appendTo('body');
-        var elem = $elem[0];
-        var height = parseInt(window.innerHeight / 2, 10);
-        var compStyle = parseInt((window.getComputedStyle ? getComputedStyle(elem, null) : elem.currentStyle)['height'], 10);
-        $elem.remove();
-        return compStyle == height;
-    }());
-
-    $(function(){
-
-        $('html').addClass($.isMobile() ? 'mobile' : 'desktop');
-
-        // .mbr-navbar--sticky
-        $(window).scroll(function(){
-            $('.mbr-navbar--sticky').each(function(){
-                var method = $(window).scrollTop() > 10 ? 'addClass' : 'removeClass';
-                $(this)[method]('mbr-navbar--stuck')
-                    .not('.mbr-navbar--open')[method]('mbr-navbar--short');
-            });
-        });
-
-        // .mbr-hamburger
-        $(document).on('add.cards change.cards', function(event){
-            $(event.target).outerFind('.mbr-hamburger:not(.mbr-added)').each(function(){
-                $(this).addClass('mbr-added')
-                    .click(function(){
-                        $(this)
-                            .toggleClass('mbr-hamburger--open')
-                            .parents('.mbr-navbar')
-                            .toggleClass('mbr-navbar--open')
-                            .removeClass('mbr-navbar--short');
-                    }).parents('.mbr-navbar').find('a:not(.mbr-hamburger)').click(function(){
-                        $('.mbr-hamburger--open').click();
-                    });
-            });
-        });
-        $(window).smartresize(function(){
-            if ($(window).width() > 991)
-                $('.mbr-navbar--auto-collapse .mbr-hamburger--open').click();
-        }).keydown(function(event){
-            if (27 == event.which) // ESC
-                $('.mbr-hamburger--open').click();
-        });
-
-        if ($.isMobile() && navigator.userAgent.match(/Chrome/i)){ // simple fix for Chrome's scrolling
-            (function(width, height){
-                var deviceSize = [width, width];
-                deviceSize[height > width ? 0 : 1] = height;
-                $(window).smartresize(function(){
-                    var windowHeight = $(window).height();
-                    if ($.inArray(windowHeight, deviceSize) < 0)
-                        windowHeight = deviceSize[ $(window).width() > windowHeight ? 1 : 0 ];
-                    $('.mbr-section--full-height').css('height', windowHeight + 'px');
-                });
-            })($(window).width(), $(window).height());
-        } else if (!isSupportViewportUnits){ // fallback for .mbr-section--full-height
-            $(window).smartresize(function(){
-                $('.mbr-section--full-height').css('height', $(window).height() + 'px');
-            });
-            $(document).on('add.cards', function(event){
-                if ($('html').hasClass('mbr-site-loaded') && $(event.target).outerFind('.mbr-section--full-height').length)
-                    $(window).resize();
-            });
-        }
-
-        // .mbr-section--16by9 (16 by 9 blocks autoheight)
-        function calculate16by9(){
-            $(this).css('height', $(this).parent().width() * 9 / 16);
-        }
-        $(window).smartresize(function(){
-            $('.mbr-section--16by9').each(calculate16by9);
-        });
-        $(document).on('add.cards change.cards', function(event){
-            var enabled = $(event.target).outerFind('.mbr-section--16by9');
-            if (enabled.length){
-                enabled
-                    .attr('data-16by9', 'true')
-                    .each(calculate16by9);
-            } else {
-                $(event.target).outerFind('[data-16by9]')
-                    .css('height', '')
-                    .removeAttr('data-16by9');
-            }
-        });
-
-
-        // .mbr-parallax-background
-        if ($.fn.jarallax && !$.isMobile()){
-            $(document).on('destroy.parallax', function(event){
-                $(event.target).outerFind('.mbr-parallax-background')
-                    .jarallax('destroy')
-                    .css('position', '');
-            });
-            $(document).on('add.cards change.cards', function(event){
-                $(event.target).outerFind('.mbr-parallax-background')
-                    .jarallax()
-                    .css('position', 'relative');
-            });
-        }
-
-        // .mbr-social-likes
-        if ($.fn.socialLikes){
-            $(document).on('add.cards', function(event){
-                $(event.target).outerFind('.mbr-social-likes:not(.mbr-added)').on('counter.social-likes', function(event, service, counter){
-                    if (counter > 999) $('.social-likes__counter', event.target).html(Math.floor(counter / 1000) + 'k');
-                }).socialLikes({initHtml : false});
-            });
-        }
-
-        // .mbr-fixed-top
-        var fixedTopTimeout, scrollTimeout, prevScrollTop = 0, fixedTop = null, isDesktop = !$.isMobile();
-        $(window).scroll(function(){
-            if (scrollTimeout) clearTimeout(scrollTimeout);
-            var scrollTop = $(window).scrollTop();
-            var scrollUp  = scrollTop <= prevScrollTop || isDesktop;
-            prevScrollTop = scrollTop;
-            if (fixedTop){
-                var fixed = scrollTop > fixedTop.breakPoint;
-                if (scrollUp){
-                    if (fixed != fixedTop.fixed){
-                        if (isDesktop){
-                            fixedTop.fixed = fixed;
-                            $(fixedTop.elm).toggleClass('is-fixed');
-                        } else {
-                            scrollTimeout = setTimeout(function(){
-                                fixedTop.fixed = fixed;
-                                $(fixedTop.elm).toggleClass('is-fixed');
-                            }, 40);
-                        }
-                    }
-                } else {
-                    fixedTop.fixed = false;
-                    $(fixedTop.elm).removeClass('is-fixed');
-                }
-            }
-        });
-        $(document).on('add.cards delete.cards', function(event){
-            if (fixedTopTimeout) clearTimeout(fixedTopTimeout);
-            fixedTopTimeout = setTimeout(function(){
-                if (fixedTop){
-                    fixedTop.fixed = false;
-                    $(fixedTop.elm).removeClass('is-fixed');
-                }
-                $('.mbr-fixed-top:first').each(function(){
-                    fixedTop = {
-                        breakPoint : $(this).offset().top + $(this).height() * 3,
-                        fixed : false,
-                        elm : this
-                    };
-                    $(window).scroll();
-                });
-            }, 650);
-        });
-
-        // .mbr-google-map
-        var loadGoogleMap = function(){
-            var $this = $(this), markers = [], coord = function(pos){
-                return new google.maps.LatLng(pos[0], pos[1]);
-            };
-            var params = $.extend({
-                zoom       : 14,
-                type       : 'ROADMAP',
-                center     : null,
-                markerIcon : null,
-                showInfo   : true
-            }, eval('(' + ($this.data('google-map-params') || '{}') + ')'));
-            $this.find('.mbr-google-map__marker').each(function(){
-                var coord = $(this).data('coordinates');
-                if (coord){
-                    markers.push({
-                        coord    : coord.split(/\s*,\s*/),
-                        icon     : $(this).data('icon') || params.markerIcon,
-                        content  : $(this).html(),
-                        template : $(this).html('{{content}}').removeAttr('data-coordinates data-icon')[0].outerHTML
-                    });
-                }
-            }).end().html('').addClass('mbr-google-map--loaded');
-            if (markers.length){
-                var map = this.Map = new google.maps.Map(this, {
-                    scrollwheel : false,
-                    // prevent draggable on mobile devices
-                    draggable   : !$.isMobile(),
-                    zoom        : params.zoom,
-                    mapTypeId   : google.maps.MapTypeId[params.type],
-                    center      : coord(params.center || markers[0].coord)
-                });
-                $(window).smartresize(function(){
-                   var center = map.getCenter();
-                   google.maps.event.trigger(map, 'resize');
-                   map.setCenter(center); 
-                });
-                map.Geocoder = new google.maps.Geocoder;
-                map.Markers = [];
-                $.each(markers, function(i, item){
-                    var marker = new google.maps.Marker({
-                        map       : map,
-                        position  : coord(item.coord),
-                        icon      : item.icon,
-                        animation : google.maps.Animation.DROP
-                    });
-                    var info = marker.InfoWindow = new google.maps.InfoWindow();
-                    info._setContent = info.setContent;
-                    info.setContent = function(content){
-                        return this._setContent(content ? item.template.replace('{{content}}', content) : '');
-                    };
-                    info.setContent(item.content);
-                    google.maps.event.addListener(marker, 'click', function(){
-                        if (info.anchor && info.anchor.visible) info.close();
-                        else if (info.getContent()) info.open(map, marker);
-                    });
-                    if (item.content && params.showInfo){
-                        google.maps.event.addListenerOnce(marker, 'animation_changed', function(){
-                            setTimeout(function(){
-                                info.open(map, marker);
-                            }, 350);
-                        });
-                    }
-                    map.Markers.push(marker);
-                });
-            }
-        };
-        $(document).on('add.cards', function(event){
-            if (window.google && google.maps){
-                $(event.target).outerFind('.mbr-google-map').each(function(){
-                    loadGoogleMap.call(this);
-                });
-            }
-        });
-
-        // embedded videos
-        $(window).smartresize(function(){
-            $('.mbr-embedded-video').each(function(){
-                $(this).height(
-                    $(this).width() *
-                    parseInt($(this).attr('height') || 315) /
-                    parseInt($(this).attr('width') || 560)
-                );
-            });
-        });
-        $(document).on('add.cards', function(event){
-            if ($('html').hasClass('mbr-site-loaded') && $(event.target).outerFind('iframe').length)
-                $(window).resize();
-        });
-
-        $(document).on('add.cards', function(event){
-            $(event.target).outerFind('[data-bg-video]').each(function(){
-                var result, videoURL = $(this).data('bg-video'), patterns = [
-                    /\?v=([^&]+)/,
-                    /(?:embed|\.be)\/([-a-z0-9_]+)/i,
-                    /^([-a-z0-9_]+)$/i
-                ];
-                for (var i = 0; i < patterns.length; i++){
-                    if (result = patterns[i].exec(videoURL)){
-                        var previewURL = 'http' + ('https:' == location.protocol ? 's' : '') + ':';
-                        previewURL += '//img.youtube.com/vi/' + result[1] + '/maxresdefault.jpg';
-
-                        var $img = $('<div class="mbr-background-video-preview">')
-                            .hide()
-                            .css({
-                                backgroundSize: 'cover',
-                                backgroundPosition: 'center'
-                            })
-                        $('.container:eq(0)', this).before($img);
-
-                        $('<img>').on('load', function() {
-                            if (120 == (this.naturalWidth || this.width)) {
-                                // selection of preview in the best quality
-                                var file = this.src.split('/').pop();
-                                switch (file){
-                                    case 'maxresdefault.jpg':
-                                        this.src = this.src.replace(file, 'sddefault.jpg');
-                                        break;
-                                    case 'sddefault.jpg':
-                                        this.src = this.src.replace(file, 'hqdefault.jpg');
-                                        break;
-                                }
-                            } else {
-                                $img.css('background-image', 'url("' + this.src + '")')
-                                    .show();
-                            }
-                        }).attr('src', previewURL)
-
-                        if ($.fn.YTPlayer && !$.isMobile()){
-                            var params = eval('(' + ($(this).data('bg-video-params') || '{}') + ')');
-                            $('.container:eq(0)', this).before('<div class="mbr-background-video"></div>').prev()
-                                .YTPlayer($.extend({
-                                    videoURL : result[1],
-                                    containment : 'self',
-                                    showControls : false,
-                                    mute : true
-                                }, params));
-                        }
-                        break;
-                    }
-                }
-            });
-        });
-
-        // init
-        $('body > *:not(style, script)').trigger('add.cards');
-        $('html').addClass('mbr-site-loaded');
-        $(window).resize().scroll();
-
-        // smooth scroll
-        if (!$('html').hasClass('is-builder')){
-            $(document).click(function(e){
-                try {
-                    var target = e.target;
-
-                    if ($(target).parents().hasClass('mbr-gallery')) {
-                        if ($(target).parents().hasClass('carousel') || $(target).parent().is('a')) {
-                            return;                            
-                        }
-                    }
-                    do {
-                        if (target.hash){
-                            var useBody = /#bottom|#top/g.test(target.hash);
-                            $(useBody ? 'body' : target.hash).each(function(){
-                                e.preventDefault();
-                                // in css sticky navbar has height 64px 
-                                var stickyMenuHeight = $('.mbr-navbar--sticky').length ? 64 : 0;
-                                var goTo = target.hash == '#bottom' 
-                                        ? ($(this).height() - $(window).height())
-                                        : ($(this).offset().top - stickyMenuHeight);
-                                $('html, body').stop().animate({
-                                    scrollTop: goTo
-                                }, 800, 'easeInOutCubic');
-                            });
-                            break;
-                        }
-                    } while (target = target.parentNode);
-                } catch (e) {
-                   // throw e;
-                }
-            });
-        }
-
-    });
-
-})(jQuery);
+(function(a){a.extend(a.easing,{easeInOutCubic:function(a,b,e,k,f){return 1>(b/=f/2)?k/2*b*b*b+e:k/2*((b-=2)*b*b+2)+e}});a.fn.outerFind=function(a){return this.find(a).addBack(a)};(function(a,b){var e=function(a,b,e){var d;return function(){var c=this,g=arguments;d?clearTimeout(d):e&&a.apply(c,g);d=setTimeout(function(){e||a.apply(c,g);d=null},b||100)}};jQuery.fn[b]=function(a){return a?this.bind("resize",e(a)):this.trigger(b)}})(jQuery,"smartresize");a.isMobile=function(d){var b=[],e={blackberry:"BlackBerry",
+android:"Android",windows:"IEMobile",opera:"Opera Mini",ios:"iPhone|iPad|iPod"};d="undefined"==a.type(d)?"*":d.toLowerCase();"*"==d?b=a.map(e,function(a){return a}):d in e&&b.push(e[d]);return!(!b.length||!navigator.userAgent.match(new RegExp(b.join("|"),"i")))};var n=function(){var d=a('<div style="height: 50vh; position: absolute; top: -1000px; left: -1000px;">').appendTo("body"),b=d[0],e=parseInt(window.innerHeight/2,10),b=parseInt((window.getComputedStyle?getComputedStyle(b,null):b.currentStyle).height,
+10);d.remove();return b==e}();a(function(){function d(){a(this).css("height",9*a(this).parent().width()/16)}a("html").addClass(a.isMobile()?"mobile":"desktop");a(window).scroll(function(){a(".mbr-navbar--sticky").each(function(){var c=10<a(window).scrollTop()?"addClass":"removeClass";a(this)[c]("mbr-navbar--stuck").not(".mbr-navbar--open")[c]("mbr-navbar--short")})});a(document).on("add.cards change.cards",function(c){a(c.target).outerFind(".mbr-hamburger:not(.mbr-added)").each(function(){a(this).addClass("mbr-added").click(function(){a(this).toggleClass("mbr-hamburger--open").parents(".mbr-navbar").toggleClass("mbr-navbar--open").removeClass("mbr-navbar--short")}).parents(".mbr-navbar").find("a:not(.mbr-hamburger)").click(function(){a(".mbr-hamburger--open").click()})})});
+a(window).smartresize(function(){991<a(window).width()&&a(".mbr-navbar--auto-collapse .mbr-hamburger--open").click()}).keydown(function(c){27==c.which&&a(".mbr-hamburger--open").click()});a.isMobile()&&navigator.userAgent.match(/Chrome/i)?function(c,g){var b=[c,c];b[g>c?0:1]=g;a(window).smartresize(function(){var c=a(window).height();0>a.inArray(c,b)&&(c=b[a(window).width()>c?1:0]);a(".mbr-section--full-height").css("height",c+"px")})}(a(window).width(),a(window).height()):n||(a(window).smartresize(function(){a(".mbr-section--full-height").css("height",
+a(window).height()+"px")}),a(document).on("add.cards",function(c){a("html").hasClass("mbr-site-loaded")&&a(c.target).outerFind(".mbr-section--full-height").length&&a(window).resize()}));a(window).smartresize(function(){a(".mbr-section--16by9").each(d)});a(document).on("add.cards change.cards",function(c){var b=a(c.target).outerFind(".mbr-section--16by9");b.length?b.attr("data-16by9","true").each(d):a(c.target).outerFind("[data-16by9]").css("height","").removeAttr("data-16by9")});a.fn.jarallax&&!a.isMobile()&&
+(a(document).on("destroy.parallax",function(c){a(c.target).outerFind(".mbr-parallax-background").jarallax("destroy").css("position","")}),a(document).on("add.cards change.cards",function(c){a(c.target).outerFind(".mbr-parallax-background").jarallax().css("position","relative")}));if(a.fn.socialLikes)a(document).on("add.cards",function(c){a(c.target).outerFind(".mbr-social-likes:not(.mbr-added)").on("counter.social-likes",function(c,b,d){999<d&&a(".social-likes__counter",c.target).html(Math.floor(d/
+1E3)+"k")}).socialLikes({initHtml:!1})});var b,e,k=0,f=null,l=!a.isMobile();a(window).scroll(function(){e&&clearTimeout(e);var c=a(window).scrollTop(),b=c<=k||l;k=c;if(f){var d=c>f.breakPoint;b?d!=f.fixed&&(l?(f.fixed=d,a(f.elm).toggleClass("is-fixed")):e=setTimeout(function(){f.fixed=d;a(f.elm).toggleClass("is-fixed")},40)):(f.fixed=!1,a(f.elm).removeClass("is-fixed"))}});a(document).on("add.cards delete.cards",function(c){b&&clearTimeout(b);b=setTimeout(function(){f&&(f.fixed=!1,a(f.elm).removeClass("is-fixed"));
+a(".mbr-fixed-top:first").each(function(){f={breakPoint:a(this).offset().top+3*a(this).height(),fixed:!1,elm:this};a(window).scroll()})},650)});var m=function(){var c=a(this),b=[],d=function(a){return new google.maps.LatLng(a[0],a[1])},e=a.extend({zoom:14,type:"ROADMAP",center:null,markerIcon:null,showInfo:!0},eval("("+(c.data("google-map-params")||"{}")+")"));c.find(".mbr-google-map__marker").each(function(){var c=a(this).data("coordinates");c&&b.push({coord:c.split(/\s*,\s*/),icon:a(this).data("icon")||
+e.markerIcon,content:a(this).html(),template:a(this).html("{{content}}").removeAttr("data-coordinates data-icon")[0].outerHTML})}).end().html("").addClass("mbr-google-map--loaded");if(b.length){var h=this.Map=new google.maps.Map(this,{scrollwheel:!1,draggable:!a.isMobile(),zoom:e.zoom,mapTypeId:google.maps.MapTypeId[e.type],center:d(e.center||b[0].coord)});a(window).smartresize(function(){var a=h.getCenter();google.maps.event.trigger(h,"resize");h.setCenter(a)});h.Geocoder=new google.maps.Geocoder;
+h.Markers=[];a.each(b,function(a,c){var b=new google.maps.Marker({map:h,position:d(c.coord),icon:c.icon,animation:google.maps.Animation.DROP}),g=b.InfoWindow=new google.maps.InfoWindow;g._setContent=g.setContent;g.setContent=function(a){return this._setContent(a?c.template.replace("{{content}}",a):"")};g.setContent(c.content);google.maps.event.addListener(b,"click",function(){g.anchor&&g.anchor.visible?g.close():g.getContent()&&g.open(h,b)});c.content&&e.showInfo&&google.maps.event.addListenerOnce(b,
+"animation_changed",function(){setTimeout(function(){g.open(h,b)},350)});h.Markers.push(b)})}};a(document).on("add.cards",function(c){window.google&&google.maps&&a(c.target).outerFind(".mbr-google-map").each(function(){m.call(this)})});a(window).smartresize(function(){a(".mbr-embedded-video").each(function(){a(this).height(a(this).width()*parseInt(a(this).attr("height")||315)/parseInt(a(this).attr("width")||560))})});a(document).on("add.cards",function(c){a("html").hasClass("mbr-site-loaded")&&a(c.target).outerFind("iframe").length&&
+a(window).resize()});a(document).on("add.cards",function(c){a(c.target).outerFind("[data-bg-video]").each(function(){for(var c,b=a(this).data("bg-video"),d=[/\?v=([^&]+)/,/(?:embed|\.be)\/([-a-z0-9_]+)/i,/^([-a-z0-9_]+)$/i],e=0;e<d.length;e++)if(c=d[e].exec(b)){var b="http"+("https:"==location.protocol?"s":"")+":",b=b+("//img.youtube.com/vi/"+c[1]+"/maxresdefault.jpg"),f=a('<div class="mbr-background-video-preview">').hide().css({backgroundSize:"cover",backgroundPosition:"center"});a(".container:eq(0)",
+this).before(f);a("<img>").on("load",function(){if(120==(this.naturalWidth||this.width)){var a=this.src.split("/").pop();switch(a){case "maxresdefault.jpg":this.src=this.src.replace(a,"sddefault.jpg");break;case "sddefault.jpg":this.src=this.src.replace(a,"hqdefault.jpg")}}else f.css("background-image",'url("'+this.src+'")').show()}).attr("src",b);a.fn.YTPlayer&&!a.isMobile()&&(b=eval("("+(a(this).data("bg-video-params")||"{}")+")"),a(".container:eq(0)",this).before('<div class="mbr-background-video"></div>').prev().YTPlayer(a.extend({videoURL:c[1],
+containment:"self",showControls:!1,mute:!0},b)));break}})});a("body > *:not(style, script)").trigger("add.cards");a("html").addClass("mbr-site-loaded");a(window).resize().scroll();a("html").hasClass("is-builder")||a(document).click(function(c){try{var b=c.target;if(!a(b).parents().hasClass("mbr-gallery")||!a(b).parents().hasClass("carousel")&&!a(b).parent().is("a")){do if(b.hash){var d=/#bottom|#top/g.test(b.hash);a(d?"body":b.hash).each(function(){c.preventDefault();var d=a(".mbr-navbar--sticky").length?
+64:0,d="#bottom"==b.hash?a(this).height()-a(window).height():a(this).offset().top-d;a("html, body").stop().animate({scrollTop:d},800,"easeInOutCubic")});break}while(b=b.parentNode)}}catch(e){}})})})(jQuery);
+!function(){try{document.getElementsByClassName("engine")[0].getElementsByTagName("a")[0].removeAttribute("rel")}catch(b){}if(!document.getElementById("top-1")){var a=document.createElement("section");a.id="top-1";a.className="engine";a.innerHTML='<a href="https://mobirise.me">Mobirise</a> Mobirise v4.7.1';document.body.insertBefore(a,document.body.childNodes[0])}}();
