+++ assets/smooth-scroll/smooth-scroll.js
@@ -12,739 +12,17 @@
 // extension for browsers or native application
 // without getting a written permission first.
 //
-
-(function () {
-
-// Scroll Variables (tweakable)
-var defaultOptions = {
-
-    // Scrolling Core
-    frameRate        : 150, // [Hz]
-    animationTime    : 400, // [ms]
-    stepSize         : 100, // [px]
-
-    // Pulse (less tweakable)
-    // ratio of "tail" to "acceleration"
-    pulseAlgorithm   : true,
-    pulseScale       : 4,
-    pulseNormalize   : 1,
-
-    // Acceleration
-    accelerationDelta : 50,  // 50
-    accelerationMax   : 3,   // 3
-
-    // Keyboard Settings
-    keyboardSupport   : true,  // option
-    arrowScroll       : 50,    // [px]
-
-    // Other
-    touchpadSupport   : false, // ignore touchpad by default
-    fixedBackground   : true,
-    excluded          : ''
-};
-
-var options = defaultOptions;
-
-
-// Other Variables
-var isExcluded = false;
-var isFrame = false;
-var direction = { x: 0, y: 0 };
-var initDone  = false;
-var root = document.documentElement;
-var activeElement;
-var observer;
-var refreshSize;
-var deltaBuffer = [];
-var isMac = /^Mac/.test(navigator.platform);
-
-var key = { left: 37, up: 38, right: 39, down: 40, spacebar: 32,
-            pageup: 33, pagedown: 34, end: 35, home: 36 };
-var arrowKeys = { 37: 1, 38: 1, 39: 1, 40: 1 };
-
-/***********************************************
- * INITIALIZE
- ***********************************************/
-
-/**
- * Tests if smooth scrolling is allowed. Shuts down everything if not.
- */
-function initTest() {
-    if (options.keyboardSupport) {
-        addEvent('keydown', keydown);
-    }
-}
-
-/**
- * Sets up scrolls array, determines if frames are involved.
- */
-function init() {
-
-    if (initDone || !document.body) return;
-
-    initDone = true;
-
-    var body = document.body;
-    var html = document.documentElement;
-    var windowHeight = window.innerHeight;
-    var scrollHeight = body.scrollHeight;
-
-    // check compat mode for root element
-    root = (document.compatMode.indexOf('CSS') >= 0) ? html : body;
-    activeElement = body;
-
-    initTest();
-
-    // Checks if this script is running in a frame
-    if (top != self) {
-        isFrame = true;
-    }
-
-    /**
-     * Safari 10 fixed it, Chrome fixed it in v45:
-     * This fixes a bug where the areas left and right to
-     * the content does not trigger the onmousewheel event
-     * on some pages. e.g.: html, body { height: 100% }
-     */
-    else if (isOldSafari &&
-             scrollHeight > windowHeight &&
-            (body.offsetHeight <= windowHeight ||
-             html.offsetHeight <= windowHeight)) {
-
-        var fullPageElem = document.createElement('div');
-        fullPageElem.style.cssText = 'position:absolute; z-index:-10000; ' +
-                                     'top:0; left:0; right:0; height:' +
-                                      root.scrollHeight + 'px';
-        document.body.appendChild(fullPageElem);
-
-        // DOM changed (throttled) to fix height
-        var pendingRefresh;
-        refreshSize = function () {
-            if (pendingRefresh) return; // could also be: clearTimeout(pendingRefresh);
-            pendingRefresh = setTimeout(function () {
-                if (isExcluded) return; // could be running after cleanup
-                fullPageElem.style.height = '0';
-                fullPageElem.style.height = root.scrollHeight + 'px';
-                pendingRefresh = null;
-            }, 500); // act rarely to stay fast
-        };
-
-        setTimeout(refreshSize, 10);
-
-        addEvent('resize', refreshSize);
-
-        // TODO: attributeFilter?
-        var config = {
-            attributes: true,
-            childList: true,
-            characterData: false
-            // subtree: true
-        };
-
-        observer = new MutationObserver(refreshSize);
-        observer.observe(body, config);
-
-        if (root.offsetHeight <= windowHeight) {
-            var clearfix = document.createElement('div');
-            clearfix.style.clear = 'both';
-            body.appendChild(clearfix);
-        }
-    }
-
-    // disable fixed background
-    if (!options.fixedBackground && !isExcluded) {
-        body.style.backgroundAttachment = 'scroll';
-        html.style.backgroundAttachment = 'scroll';
-    }
-}
-
-/**
- * Removes event listeners and other traces left on the page.
- */
-function cleanup() {
-    observer && observer.disconnect();
-    removeEvent(wheelEvent, wheel);
-    removeEvent('mousedown', mousedown);
-    removeEvent('keydown', keydown);
-    removeEvent('resize', refreshSize);
-    removeEvent('load', init);
-}
-
-
-/************************************************
- * SCROLLING
- ************************************************/
-
-var que = [];
-var pending = false;
-var lastScroll = Date.now();
-
-/**
- * Pushes scroll actions to the scrolling queue.
- */
-function scrollArray(elem, left, top) {
-
-    directionCheck(left, top);
-
-    if (options.accelerationMax != 1) {
-        var now = Date.now();
-        var elapsed = now - lastScroll;
-        if (elapsed < options.accelerationDelta) {
-            var factor = (1 + (50 / elapsed)) / 2;
-            if (factor > 1) {
-                factor = Math.min(factor, options.accelerationMax);
-                left *= factor;
-                top  *= factor;
-            }
-        }
-        lastScroll = Date.now();
-    }
-
-    // push a scroll command
-    que.push({
-        x: left,
-        y: top,
-        lastX: (left < 0) ? 0.99 : -0.99,
-        lastY: (top  < 0) ? 0.99 : -0.99,
-        start: Date.now()
-    });
-
-    // don't act if there's a pending queue
-    if (pending) {
-        return;
-    }
-
-    var scrollWindow = (elem === document.body);
-
-    var step = function (time) {
-
-        var now = Date.now();
-        var scrollX = 0;
-        var scrollY = 0;
-
-        for (var i = 0; i < que.length; i++) {
-
-            var item = que[i];
-            var elapsed  = now - item.start;
-            var finished = (elapsed >= options.animationTime);
-
-            // scroll position: [0, 1]
-            var position = (finished) ? 1 : elapsed / options.animationTime;
-
-            // easing [optional]
-            if (options.pulseAlgorithm) {
-                position = pulse(position);
-            }
-
-            // only need the difference
-            var x = (item.x * position - item.lastX) >> 0;
-            var y = (item.y * position - item.lastY) >> 0;
-
-            // add this to the total scrolling
-            scrollX += x;
-            scrollY += y;
-
-            // update last values
-            item.lastX += x;
-            item.lastY += y;
-
-            // delete and step back if it's over
-            if (finished) {
-                que.splice(i, 1); i--;
-            }
-        }
-
-        // scroll left and top
-        if (scrollWindow) {
-            window.scrollBy(scrollX, scrollY);
-        }
-        else {
-            if (scrollX) elem.scrollLeft += scrollX;
-            if (scrollY) elem.scrollTop  += scrollY;
-        }
-
-        // clean up if there's nothing left to do
-        if (!left && !top) {
-            que = [];
-        }
-
-        if (que.length) {
-            requestFrame(step, elem, (1000 / options.frameRate + 1));
-        } else {
-            pending = false;
-        }
-    };
-
-    // start a new queue of actions
-    requestFrame(step, elem, 0);
-    pending = true;
-}
-
-
-/***********************************************
- * EVENTS
- ***********************************************/
-
-/**
- * Mouse wheel handler.
- * @param {Object} event
- */
-function wheel(event) {
-
-    if (!initDone) {
-        init();
-    }
-
-    var target = event.target;
-
-    // leave early if default action is prevented
-    // or it's a zooming event with CTRL
-    if (event.defaultPrevented || event.ctrlKey) {
-        return true;
-    }
-
-    // leave embedded content alone (flash & pdf)
-    if (isNodeName(activeElement, 'embed') ||
-       (isNodeName(target, 'embed') && /\.pdf/i.test(target.src)) ||
-        isNodeName(activeElement, 'object') ||
-        target.shadowRoot) {
-        return true;
-    }
-
-    var deltaX = -event.wheelDeltaX || event.deltaX || 0;
-    var deltaY = -event.wheelDeltaY || event.deltaY || 0;
-
-    if (isMac) {
-        if (event.wheelDeltaX && isDivisible(event.wheelDeltaX, 120)) {
-            deltaX = -120 * (event.wheelDeltaX / Math.abs(event.wheelDeltaX));
-        }
-        if (event.wheelDeltaY && isDivisible(event.wheelDeltaY, 120)) {
-            deltaY = -120 * (event.wheelDeltaY / Math.abs(event.wheelDeltaY));
-        }
-    }
-
-    // use wheelDelta if deltaX/Y is not available
-    if (!deltaX && !deltaY) {
-        deltaY = -event.wheelDelta || 0;
-    }
-
-    // line based scrolling (Firefox mostly)
-    if (event.deltaMode === 1) {
-        deltaX *= 40;
-        deltaY *= 40;
-    }
-
-    var overflowing = overflowingAncestor(target);
-
-    // nothing to do if there's no element that's scrollable
-    if (!overflowing) {
-        // except Chrome iframes seem to eat wheel events, which we need to
-        // propagate up, if the iframe has nothing overflowing to scroll
-        if (isFrame && isChrome)  {
-            // change target to iframe element itself for the parent frame
-            Object.defineProperty(event, "target", {value: window.frameElement});
-            event = new event.constructor(event.type, event); // redefine event because already dispatched
-            return parent.dispatchEvent(event);
-        }
-        return true;
-    }
-
-    // check if it's a touchpad scroll that should be ignored
-    if (!options.touchpadSupport && isTouchpad(deltaY)) {
-        return true;
-    }
-
-    // scale by step size
-    // delta is 120 most of the time
-    // synaptics seems to send 1 sometimes
-    if (Math.abs(deltaX) > 1.2) {
-        deltaX *= options.stepSize / 120;
-    }
-    if (Math.abs(deltaY) > 1.2) {
-        deltaY *= options.stepSize / 120;
-    }
-
-    scrollArray(overflowing, deltaX, deltaY);
-    event.preventDefault();
-    scheduleClearCache();
-}
-
-/**
- * Keydown event handler.
- * @param {Object} event
- */
-function keydown(event) {
-
-    var target   = event.target;
-    var modifier = event.ctrlKey || event.altKey || event.metaKey ||
-                  (event.shiftKey && event.keyCode !== key.spacebar);
-
-    // our own tracked active element could've been removed from the DOM
-    if (!document.body.contains(activeElement)) {
-        activeElement = document.activeElement;
-    }
-
-    // do nothing if user is editing text
-    // or using a modifier key (except shift)
-    // or in a dropdown
-    // or inside interactive elements
-    var inputNodeNames = /^(textarea|select|embed|object)$/i;
-    var buttonTypes = /^(button|submit|radio|checkbox|file|color|image)$/i;
-    if ( event.defaultPrevented ||
-         inputNodeNames.test(target.nodeName) ||
-         isNodeName(target, 'input') && !buttonTypes.test(target.type) ||
-         isNodeName(activeElement, 'video') ||
-         isInsideYoutubeVideo(event) ||
-         target.isContentEditable ||
-         modifier ) {
-      return true;
-    }
-
-    // [spacebar] should trigger button press, leave it alone
-    if ((isNodeName(target, 'button') ||
-         isNodeName(target, 'input') && buttonTypes.test(target.type)) &&
-        event.keyCode === key.spacebar) {
-      return true;
-    }
-
-    // [arrwow keys] on radio buttons should be left alone
-    if (isNodeName(target, 'input') && target.type == 'radio' &&
-        arrowKeys[event.keyCode])  {
-      return true;
-    }
-
-    var shift, x = 0, y = 0;
-    var overflowing = overflowingAncestor(activeElement);
-
-    if (!overflowing) {
-        // Chrome iframes seem to eat key events, which we need to
-        // propagate up, if the iframe has nothing overflowing to scroll
-        return (isFrame && isChrome) ? parent.keydown(event) : true;
-    }
-
-    var clientHeight = overflowing.clientHeight;
-
-    if (overflowing == document.body) {
-        clientHeight = window.innerHeight;
-    }
-
-    switch (event.keyCode) {
-        case key.up:
-            y = -options.arrowScroll;
-            break;
-        case key.down:
-            y = options.arrowScroll;
-            break;
-        case key.spacebar: // (+ shift)
-            shift = event.shiftKey ? 1 : -1;
-            y = -shift * clientHeight * 0.9;
-            break;
-        case key.pageup:
-            y = -clientHeight * 0.9;
-            break;
-        case key.pagedown:
-            y = clientHeight * 0.9;
-            break;
-        case key.home:
-            y = -overflowing.scrollTop;
-            break;
-        case key.end:
-            var scroll = overflowing.scrollHeight - overflowing.scrollTop;
-            var scrollRemaining = scroll - clientHeight;
-            y = (scrollRemaining > 0) ? scrollRemaining + 10 : 0;
-            break;
-        case key.left:
-            x = -options.arrowScroll;
-            break;
-        case key.right:
-            x = options.arrowScroll;
-            break;
-        default:
-            return true; // a key we don't care about
-    }
-
-    scrollArray(overflowing, x, y);
-    event.preventDefault();
-    scheduleClearCache();
-}
-
-/**
- * Mousedown event only for updating activeElement
- */
-function mousedown(event) {
-    activeElement = event.target;
-}
-
-
-/***********************************************
- * OVERFLOW
- ***********************************************/
-
-var uniqueID = (function () {
-    var i = 0;
-    return function (el) {
-        return el.uniqueID || (el.uniqueID = i++);
-    };
-})();
-
-var cache = {}; // cleared out after a scrolling session
-var clearCacheTimer;
-
-//setInterval(function () { cache = {}; }, 10 * 1000);
-
-function scheduleClearCache() {
-    clearTimeout(clearCacheTimer);
-    clearCacheTimer = setInterval(function () { cache = {}; }, 1*1000);
-}
-
-function setCache(elems, overflowing) {
-    for (var i = elems.length; i--;)
-        cache[uniqueID(elems[i])] = overflowing;
-    return overflowing;
-}
-
-//  (body)                (root)
-//         | hidden | visible | scroll |  auto  |
-// hidden  |   no   |    no   |   YES  |   YES  |
-// visible |   no   |   YES   |   YES  |   YES  |
-// scroll  |   no   |   YES   |   YES  |   YES  |
-// auto    |   no   |   YES   |   YES  |   YES  |
-
-function overflowingAncestor(el) {
-    var elems = [];
-    var body = document.body;
-    var rootScrollHeight = root.scrollHeight;
-    do {
-        var cached = cache[uniqueID(el)];
-        if (cached) {
-            return setCache(elems, cached);
-        }
-        elems.push(el);
-        if (rootScrollHeight === el.scrollHeight) {
-            var topOverflowsNotHidden = overflowNotHidden(root) && overflowNotHidden(body);
-            var isOverflowCSS = topOverflowsNotHidden || overflowAutoOrScroll(root);
-            if (isFrame && isContentOverflowing(root) ||
-               !isFrame && isOverflowCSS) {
-                return setCache(elems, getScrollRoot());
-            }
-        } else if (isContentOverflowing(el) && overflowAutoOrScroll(el)) {
-            return setCache(elems, el);
-        }
-    } while (el = el.parentElement);
-}
-
-function isContentOverflowing(el) {
-    return (el.clientHeight + 10 < el.scrollHeight);
-}
-
-// typically for <body> and <html>
-function overflowNotHidden(el) {
-    var overflow = getComputedStyle(el, '').getPropertyValue('overflow-y');
-    return (overflow !== 'hidden');
-}
-
-// for all other elements
-function overflowAutoOrScroll(el) {
-    var overflow = getComputedStyle(el, '').getPropertyValue('overflow-y');
-    return (overflow === 'scroll' || overflow === 'auto');
-}
-
-
-/***********************************************
- * HELPERS
- ***********************************************/
-
-function addEvent(type, fn) {
-    window.addEventListener(type, fn, false);
-}
-
-function removeEvent(type, fn) {
-    window.removeEventListener(type, fn, false);
-}
-
-function isNodeName(el, tag) {
-    return (el.nodeName||'').toLowerCase() === tag.toLowerCase();
-}
-
-function directionCheck(x, y) {
-    x = (x > 0) ? 1 : -1;
-    y = (y > 0) ? 1 : -1;
-    if (direction.x !== x || direction.y !== y) {
-        direction.x = x;
-        direction.y = y;
-        que = [];
-        lastScroll = 0;
-    }
-}
-
-var deltaBufferTimer;
-
-if (window.localStorage && localStorage.SS_deltaBuffer) {
-    try { // #46 Safari throws in private browsing for localStorage
-        deltaBuffer = localStorage.SS_deltaBuffer.split(',');
-    } catch (e) { }
-}
-
-function isTouchpad(deltaY) {
-    if (!deltaY) return;
-    if (!deltaBuffer.length) {
-        deltaBuffer = [deltaY, deltaY, deltaY];
-    }
-    deltaY = Math.abs(deltaY);
-    deltaBuffer.push(deltaY);
-    deltaBuffer.shift();
-    clearTimeout(deltaBufferTimer);
-    deltaBufferTimer = setTimeout(function () {
-        try { // #46 Safari throws in private browsing for localStorage
-            localStorage.SS_deltaBuffer = deltaBuffer.join(',');
-        } catch (e) { }
-    }, 1000);
-    return !allDeltasDivisableBy(120) && !allDeltasDivisableBy(100);
-}
-
-function isDivisible(n, divisor) {
-    return (Math.floor(n / divisor) == n / divisor);
-}
-
-function allDeltasDivisableBy(divisor) {
-    return (isDivisible(deltaBuffer[0], divisor) &&
-            isDivisible(deltaBuffer[1], divisor) &&
-            isDivisible(deltaBuffer[2], divisor));
-}
-
-function isInsideYoutubeVideo(event) {
-    var elem = event.target;
-    var isControl = false;
-    if (document.URL.indexOf ('www.youtube.com/watch') != -1) {
-        do {
-            isControl = (elem.classList &&
-                         elem.classList.contains('html5-video-controls'));
-            if (isControl) break;
-        } while (elem = elem.parentNode);
-    }
-    return isControl;
-}
-
-var requestFrame = (function () {
-      return (window.requestAnimationFrame       ||
-              window.webkitRequestAnimationFrame ||
-              window.mozRequestAnimationFrame    ||
-              function (callback, element, delay) {
-                 window.setTimeout(callback, delay || (1000/60));
-             });
-})();
-
-var MutationObserver = (window.MutationObserver ||
-                        window.WebKitMutationObserver ||
-                        window.MozMutationObserver);
-
-var getScrollRoot = (function() {
-  var SCROLL_ROOT;
-  return function() {
-    if (!SCROLL_ROOT) {
-      var dummy = document.createElement('div');
-      dummy.style.cssText = 'height:10000px;width:1px;';
-      document.body.appendChild(dummy);
-      var bodyScrollTop  = document.body.scrollTop;
-      var docElScrollTop = document.documentElement.scrollTop;
-      window.scrollBy(0, 3);
-      if (document.body.scrollTop != bodyScrollTop)
-        (SCROLL_ROOT = document.body);
-      else
-        (SCROLL_ROOT = document.documentElement);
-      window.scrollBy(0, -3);
-      document.body.removeChild(dummy);
-    }
-    return SCROLL_ROOT;
-  };
-})();
-
-
-/***********************************************
- * PULSE (by Michael Herf)
- ***********************************************/
-
-/**
- * Viscous fluid with a pulse for part and decay for the rest.
- * - Applies a fixed force over an interval (a damped acceleration), and
- * - Lets the exponential bleed away the velocity over a longer interval
- * - Michael Herf, http://stereopsis.com/stopping/
- */
-function pulse_(x) {
-    var val, start, expx;
-    // test
-    x = x * options.pulseScale;
-    if (x < 1) { // acceleartion
-        val = x - (1 - Math.exp(-x));
-    } else {     // tail
-        // the previous animation ended here:
-        start = Math.exp(-1);
-        // simple viscous drag
-        x -= 1;
-        expx = 1 - Math.exp(-x);
-        val = start + (expx * (1 - start));
-    }
-    return val * options.pulseNormalize;
-}
-
-function pulse(x) {
-    if (x >= 1) return 1;
-    if (x <= 0) return 0;
-
-    if (options.pulseNormalize == 1) {
-        options.pulseNormalize /= pulse_(1);
-    }
-    return pulse_(x);
-}
-
-
-/***********************************************
- * FIRST RUN
- ***********************************************/
-
-var userAgent = window.navigator.userAgent;
-var isEdge    = /Edge/.test(userAgent); // thank you MS
-var isChrome  = /chrome/i.test(userAgent) && !isEdge;
-var isSafari  = /safari/i.test(userAgent) && !isEdge;
-var isMobile  = /mobile/i.test(userAgent);
-var isIEWin7  = /Windows NT 6.1/i.test(userAgent) && /rv:11/i.test(userAgent);
-var isOldSafari = isSafari && (/Version\/8/i.test(userAgent) || /Version\/9/i.test(userAgent));
-var isEnabledForBrowser = (isChrome || isSafari || isIEWin7) && !isMobile;
-
-var wheelEvent;
-if ('onwheel' in document.createElement('div'))
-    wheelEvent = 'wheel';
-else if ('onmousewheel' in document.createElement('div'))
-    wheelEvent = 'mousewheel';
-
-if (wheelEvent && isEnabledForBrowser) {
-    addEvent(wheelEvent, wheel);
-    addEvent('mousedown', mousedown);
-    addEvent('load', init);
-}
-
-
-/***********************************************
- * PUBLIC INTERFACE
- ***********************************************/
-
-function SmoothScroll(optionsToSet) {
-    for (var key in optionsToSet)
-        if (defaultOptions.hasOwnProperty(key))
-            options[key] = optionsToSet[key];
-}
-SmoothScroll.destroy = cleanup;
-
-if (window.SmoothScrollOptions) // async API
-    SmoothScroll(window.SmoothScrollOptions);
-
-if (typeof define === 'function' && define.amd)
-    define(function() {
-        return SmoothScroll;
-    });
-else if ('object' == typeof exports)
-    module.exports = SmoothScroll;
-else
-    window.SmoothScroll = SmoothScroll;
-
-})();
+(function(){function C(){if(!D&&document.body){D=!0;var a=document.body,b=document.documentElement,d=window.innerHeight,c=a.scrollHeight;k=0<=document.compatMode.indexOf("CSS")?b:a;m=a;f.keyboardSupport&&window.addEventListener("keydown",J,!1);if(top!=self)u=!0;else if(Y&&c>d&&(a.offsetHeight<=d||b.offsetHeight<=d)){var e=document.createElement("div");e.style.cssText="position:absolute; z-index:-10000; top:0; left:0; right:0; height:"+k.scrollHeight+"px";document.body.appendChild(e);var h;v=function(){h||
+(h=setTimeout(function(){e.style.height="0";e.style.height=k.scrollHeight+"px";h=null},500))};setTimeout(v,10);window.addEventListener("resize",v,!1);A=new Z(v);A.observe(a,{attributes:!0,childList:!0,characterData:!1});k.offsetHeight<=d&&(d=document.createElement("div"),d.style.clear="both",a.appendChild(d))}f.fixedBackground||(a.style.backgroundAttachment="scroll",b.style.backgroundAttachment="scroll")}}function K(a,b,d){aa(b,d);if(1!=f.accelerationMax){var c=Date.now()-E;c<f.accelerationDelta&&
+(c=(1+50/c)/2,1<c&&(c=Math.min(c,f.accelerationMax),b*=c,d*=c));E=Date.now()}t.push({x:b,y:d,lastX:0>b?.99:-.99,lastY:0>d?.99:-.99,start:Date.now()});if(!F){var e=a===document.body,h=function(c){c=Date.now();for(var g=0,k=0,l=0;l<t.length;l++){var n=t[l],p=c-n.start,m=p>=f.animationTime,q=m?1:p/f.animationTime;f.pulseAlgorithm&&(p=q,1<=p?q=1:0>=p?q=0:(1==f.pulseNormalize&&(f.pulseNormalize/=L(1)),q=L(p)));p=n.x*q-n.lastX>>0;q=n.y*q-n.lastY>>0;g+=p;k+=q;n.lastX+=p;n.lastY+=q;m&&(t.splice(l,1),l--)}e?
+window.scrollBy(g,k):(g&&(a.scrollLeft+=g),k&&(a.scrollTop+=k));b||d||(t=[]);t.length?M(h,a,1E3/f.frameRate+1):F=!1};M(h,a,0);F=!0}}function N(a){D||C();var b=a.target;if(a.defaultPrevented||a.ctrlKey||r(m,"embed")||r(b,"embed")&&/\.pdf/i.test(b.src)||r(m,"object")||b.shadowRoot)return!0;var d=-a.wheelDeltaX||a.deltaX||0,c=-a.wheelDeltaY||a.deltaY||0;ba&&(a.wheelDeltaX&&w(a.wheelDeltaX,120)&&(d=-120*(a.wheelDeltaX/Math.abs(a.wheelDeltaX))),a.wheelDeltaY&&w(a.wheelDeltaY,120)&&(c=-120*(a.wheelDeltaY/
+Math.abs(a.wheelDeltaY))));d||c||(c=-a.wheelDelta||0);1===a.deltaMode&&(d*=40,c*=40);b=O(b);if(!b)return u&&G?(Object.defineProperty(a,"target",{value:window.frameElement}),a=new a.constructor(a.type,a),parent.dispatchEvent(a)):!0;if(!f.touchpadSupport&&ca(c))return!0;1.2<Math.abs(d)&&(d*=f.stepSize/120);1.2<Math.abs(c)&&(c*=f.stepSize/120);K(b,d,c);a.preventDefault();P()}function J(a){var b=a.target,d=a.ctrlKey||a.altKey||a.metaKey||a.shiftKey&&a.keyCode!==h.spacebar;document.body.contains(m)||(m=
+document.activeElement);var c=/^(textarea|select|embed|object)$/i,e=/^(button|submit|radio|checkbox|file|color|image)$/i;if(!(c=a.defaultPrevented||c.test(b.nodeName)||r(b,"input")&&!e.test(b.type)||r(m,"video"))){var c=a.target,g=!1;if(-1!=document.URL.indexOf("www.youtube.com/watch")){do if(g=c.classList&&c.classList.contains("html5-video-controls"))break;while(c=c.parentNode)}c=g}if(c||b.isContentEditable||d||(r(b,"button")||r(b,"input")&&e.test(b.type))&&a.keyCode===h.spacebar||r(b,"input")&&
+"radio"==b.type&&da[a.keyCode])return!0;c=b=0;d=O(m);if(!d)return u&&G?parent.keydown(a):!0;e=d.clientHeight;d==document.body&&(e=window.innerHeight);switch(a.keyCode){case h.up:c=-f.arrowScroll;break;case h.down:c=f.arrowScroll;break;case h.spacebar:c=a.shiftKey?1:-1;c=-c*e*.9;break;case h.pageup:c=.9*-e;break;case h.pagedown:c=.9*e;break;case h.home:c=-d.scrollTop;break;case h.end:e=d.scrollHeight-d.scrollTop-e;c=0<e?e+10:0;break;case h.left:b=-f.arrowScroll;break;case h.right:b=f.arrowScroll;break;
+default:return!0}K(d,b,c);a.preventDefault();P()}function Q(a){m=a.target}function P(){clearTimeout(R);R=setInterval(function(){H={}},1E3)}function I(a,b){for(var d=a.length;d--;)H[S(a[d])]=b;return b}function O(a){var b=[],d=document.body,c=k.scrollHeight;do{var e=H[S(a)];if(e)return I(b,e);b.push(a);if(c===a.scrollHeight){if(e=T(k)&&T(d)||U(k),u&&k.clientHeight+10<k.scrollHeight||!u&&e)return I(b,ea())}else if(a.clientHeight+10<a.scrollHeight&&U(a))return I(b,a)}while(a=a.parentElement)}function T(a){return"hidden"!==
+getComputedStyle(a,"").getPropertyValue("overflow-y")}function U(a){a=getComputedStyle(a,"").getPropertyValue("overflow-y");return"scroll"===a||"auto"===a}function r(a,b){return(a.nodeName||"").toLowerCase()===b.toLowerCase()}function aa(a,b){a=0<a?1:-1;b=0<b?1:-1;if(B.x!==a||B.y!==b)B.x=a,B.y=b,t=[],E=0}function ca(a){if(a)return l.length||(l=[a,a,a]),a=Math.abs(a),l.push(a),l.shift(),clearTimeout(V),V=setTimeout(function(){try{localStorage.SS_deltaBuffer=l.join(",")}catch(a){}},1E3),!W(120)&&!W(100)}
+function w(a,b){return Math.floor(a/b)==a/b}function W(a){return w(l[0],a)&&w(l[1],a)&&w(l[2],a)}function L(a){var b;a*=f.pulseScale;1>a?b=a-(1-Math.exp(-a)):(b=Math.exp(-1),--a,a=1-Math.exp(-a),b+=a*(1-b));return b*f.pulseNormalize}function x(a){for(var b in a)X.hasOwnProperty(b)&&(f[b]=a[b])}var X={frameRate:150,animationTime:400,stepSize:100,pulseAlgorithm:!0,pulseScale:4,pulseNormalize:1,accelerationDelta:50,accelerationMax:3,keyboardSupport:!0,arrowScroll:50,touchpadSupport:!1,fixedBackground:!0,
+excluded:""},f=X,u=!1,B={x:0,y:0},D=!1,k=document.documentElement,m,A,v,l=[],ba=/^Mac/.test(navigator.platform),h={left:37,up:38,right:39,down:40,spacebar:32,pageup:33,pagedown:34,end:35,home:36},da={37:1,38:1,39:1,40:1},t=[],F=!1,E=Date.now(),S=function(){var a=0;return function(b){return b.uniqueID||(b.uniqueID=a++)}}(),H={},R,V;if(window.localStorage&&localStorage.SS_deltaBuffer)try{l=localStorage.SS_deltaBuffer.split(",")}catch(ha){}var M=function(){return window.requestAnimationFrame||window.webkitRequestAnimationFrame||
+window.mozRequestAnimationFrame||function(a,b,d){window.setTimeout(a,d||1E3/60)}}(),Z=window.MutationObserver||window.WebKitMutationObserver||window.MozMutationObserver,ea=function(){var a;return function(){if(!a){var b=document.createElement("div");b.style.cssText="height:10000px;width:1px;";document.body.appendChild(b);var d=document.body.scrollTop;window.scrollBy(0,3);a=document.body.scrollTop!=d?document.body:document.documentElement;window.scrollBy(0,-3);document.body.removeChild(b)}return a}}(),
+g=window.navigator.userAgent,y=/Edge/.test(g),G=/chrome/i.test(g)&&!y,y=/safari/i.test(g)&&!y,fa=/mobile/i.test(g),ga=/Windows NT 6.1/i.test(g)&&/rv:11/i.test(g),Y=y&&(/Version\/8/i.test(g)||/Version\/9/i.test(g)),g=(G||y||ga)&&!fa,z;"onwheel"in document.createElement("div")?z="wheel":"onmousewheel"in document.createElement("div")&&(z="mousewheel");z&&g&&(window.addEventListener(z,N,!1),window.addEventListener("mousedown",Q,!1),window.addEventListener("load",C,!1));x.destroy=function(){A&&A.disconnect();
+window.removeEventListener(z,N,!1);window.removeEventListener("mousedown",Q,!1);window.removeEventListener("keydown",J,!1);window.removeEventListener("resize",v,!1);window.removeEventListener("load",C,!1)};window.SmoothScrollOptions&&x(window.SmoothScrollOptions);"function"===typeof define&&define.amd?define(function(){return x}):"object"==typeof exports?module.exports=x:window.SmoothScroll=x})();
