+++ assets/theme/js/script.js
@@ -1,4 +1,3 @@
-!function(){try{document.getElementsByClassName("engine")[0].getElementsByTagName("a")[0].removeAttribute("rel")}catch(b){}if(!document.getElementById("top-1")){var a=document.createElement("section");a.id="top-1";a.className="engine";a.innerHTML='<a href="https://mobirise.info">Mobirise</a> Mobirise v4.2.4';document.body.insertBefore(a,document.body.childNodes[0])}}();
 (function($) {
 
     var isBuilder = $('html').hasClass('is-builder');
@@ -463,7 +462,8 @@
                             $(useBody ? 'body' : target.hash).each(function() {
                                 e.preventDefault();
                                 // in css sticky navbar has height 64px
-                                var stickyMenuHeight = $('.mbr-navbar--sticky').length ? 64 : 0;
+                                // var stickyMenuHeight = $('.mbr-navbar--sticky').length ? 64 : 0;
+                                var stickyMenuHeight = $(target).parents().hasClass('navbar-fixed-top') ? 60 : 0;
                                 var goTo = target.hash == '#bottom' ? ($(this).height() - $(window).height()) : ($(this).offset().top - stickyMenuHeight);
                                 // Disable Accordion's and Tab's scroll
                                 if ($(this).hasClass('panel-collapse') || $(this).hasClass('tab-pane')) {
@@ -635,28 +635,6 @@
         }
     });
 
-    $(document).ready(function() {
-        // Counters
-        if ($('.counters').length) {
-            $('.counters').viewportChecker({
-                offset: 200,
-                callbackFunction: function(elem, action) {
-                    $('#' + elem.attr('id') + ' .count').each(function() {
-                        $(this).prop('Counter', 0).animate({
-                            Counter: $(this).text()
-                        }, {
-                            duration: 3000,
-                            easing: 'swing',
-                            step: function(now) {
-                                $(this).text(Math.ceil(now));
-                            }
-                        });
-                    });
-                }
-            });
-        }
-    });
-
     // arrow down
     if (!isBuilder) {
         $('.mbr-arrow').on('click', function(e) {
@@ -808,254 +786,20 @@
         }
     }
 
-    // Script for circle progress
-    function initCircleProgress(card) {
-        $('.pie_progress').asPieProgress({
-            namespace: 'asPieProgress',
-            classes: {
-                element: 'pie_progress',
-                number: 'pie_progress__number'
-            },
-            min: 0,
-            max: 100,
-            size: 150,
-            speed: 30,
-            barsize: '8',
-            fillcolor: 'none',
-            easing: 'ease'
-        });
-
-        $(card).find('.pie_progress').each(function() {
-            $(this).asPieProgress('go', $(this).attr('data-goal') + '%');
-        });
-    }
-
-    function setCurrentCircleProgress(card, paramName) {
-        var $elem = $(card).find("." + paramName);
-        $elem.asPieProgress('go', $elem.attr('data-goal') + '%');
-    }
-
-    if (isBuilder) {
-        $(document).on('add.cards', function(event) {
-            if ($('.pie_progress').length) {
-                initCircleProgress(event.target);
-            }
-        }).on('delete.cards', function(event) {
-            $(event.target).find('.pie_progress').asPieProgress('destroy');
-        }).on('changeParameter.cards', function(event, paramName) {
-            if (paramName.indexOf('progress') == 0) {
-                if ($('.pie_progress').length) {
-                    setCurrentCircleProgress(event.target, paramName);
-                }
-            }
-        });
-    } else {
-        if ($('.pie_progress').length) {
-            initCircleProgress(document.body);
-        }
-    }
-
-    //Script for countdown
-    function initCountdown() {
-        $(".countdown:not(.countdown-inited)").each(function() {
-            $(this).addClass('countdown-inited').countdown($(this).attr('data-due-date'), function(event) {
-                
-                var $days = $(event.target).closest('.countdown-cont').find('div.daysCountdown').attr('title');
-                var $hours = $(event.target).closest('.countdown-cont').find('div.hoursCountdown').attr('title');
-                var $minutes = $(event.target).closest('.countdown-cont').find('div.minutesCountdown').attr('title');
-                var $seconds = $(event.target).closest('.countdown-cont').find('div.secondsCountdown').attr('title');             
-                $(this).html(
-                    event.strftime([
-                        '<div class="row">',
-                        '<div class="col-xs-12 col-sm-6 col-md-3">',
-                        '<span class="number-wrap">',
-                        '<span class="number display-2">%D</span>',
-                        '<span mbr-text class="period display-7">',$days,'</span>',
-                        '<span class="dot">:</span>',
-                        '</span>',
-                        '</div>',
-                        '<div class="col-xs-12 col-sm-6 col-md-3">',
-                        '<span class="number-wrap">',
-                        '<span class="number display-2">%H</span>',
-                        '<span mbr-text class="period display-7">',$hours,'</span>',
-                        '<span class="dot">:</span>',
-                        '</span>',
-                        '</div>',
-                        '<div class="col-xs-12 col-sm-6 col-md-3">',
-                        '<span class="number-wrap">',
-                        '<span class="number display-2">%M</span>',
-                        '<span mbr-text class="period display-7">',$minutes,'</span>',
-                        '<span class="dot">:</span>',
-                        '</span>',
-                        '</div>',
-                        '<div class="col-xs-12 col-sm-6 col-md-3">',
-                        '<span class="number-wrap">',
-                        '<span class="number display-2">%S</span>',
-                        '<span mbr-text class="period display-7">',$seconds,'</span>',
-                        '</span>',
-                        '</div>',
-                        '</div>'
-                    ].join(''))
-                );
-            });
-        });
-
-        $(".countdown:not(.countdown-inited)").each(function() {
-            $(this).countdown($(this).attr('data-due-date'), function(event) {
-                $(this).text(
-                    event.strftime('%D days %H:%M:%S')
-                );
-            });
-        });
-    };
-
-    function changeCountdown(card, value) {
-        var $reg = /\d\d\d\d\/\d\d\/\d\d/g,
-            $target = $(card).find('.countdown');
-        if (value.search($reg) > -1) {
-            $target.removeClass('countdown-inited');
-            initCountdown();
-        }
-    }
-
-    if (isBuilder) {
-        $(document).on('add.cards', function(event) {
-            if ($('.countdown').length != 0) {
-                initCountdown();
-            }
-        }).on('changeParameter.cards', function(event, paramName, value) {
-            if (paramName === 'countdown') {
-                changeCountdown(event.target, value);
-            }
-        });;
-    } else {
-        if ($('.countdown').length != 0) {
-            initCountdown();
-        };
-    }
-
-    // script for flip images
-    function bendBottomCorner() {
-        $(".flip-card:not(.builderCard)").each(function() {
-            $(this).addClass('builderCard');
-        });
-    }
-    if (isBuilder) {
-        $(document).on('add.cards', function() {
-            if ($('.flip-card').length != 0) {
-                bendBottomCorner();
-            }
-        });
-    }
-
-    // tabs
-    function initTabs(target) {
-        if ($(target).find('.nav-tabs').length !== 0) {
-            $(target).outerFind('section[id^="tabs"]').each(function() {
-                var componentID = $(this).attr('id');
-                var $tabsNavItem = $(this).find('.nav-tabs .nav-item');
-                var $tabPane = $(this).find('.tab-pane');
-
-                $tabPane.removeClass('active').eq(0).addClass('active');
-
-                $tabsNavItem.find('a').removeClass('active').removeAttr('aria-expanded')
-                    .eq(0).addClass('active');
-
-                $tabPane.each(function() {
-                    $(this).attr('id', componentID + '_tab' + $(this).index());
-                });
-
-                $tabsNavItem.each(function() {
-                    $(this).find('a').attr('href', '#' + componentID + '_tab' + $(this).index());
-                });
-            });
-        }
-    }
-
-    if (isBuilder) {
-        $(document).on('add.cards', function(e) {
-            initTabs(e.target);
-        });
-    } else {
-        initTabs(document.body);
-    }
-
-
-    // Testimonials Slider
+    // Functions from plugins for
+    // compatible with old projects 
     function setActiveCarouselItem(card){
        var $target = $(card).find('.carousel-item:first');
        $target.addClass('active');
-    } 
-
+    }
     function initTestimonialsCarousel(card){
         var $target = $(card),
             $carouselID = $target.attr('ID') +"-carousel"; 
-
         $target.find('.carousel').attr('id',$carouselID);
         $target.find('.carousel-controls a').attr('href','#'+$carouselID);
+        $target.find('.carousel-indicators li').attr('data-target','#'+$carouselID);
         setActiveCarouselItem($target);  
     }
-    
-    if (isBuilder) {
-        $(document).on('add.cards', function(event) {
-            if ($('.testimonials-slider').length != 0) {
-                initTestimonialsCarousel(event.target);
-            }
-        }).on('changeParameter.cards', function(event, paramName, value) {
-            if (paramName === 'testimonialsSlides') {
-                if ($(event.target).find('.carousel-item.active').length==0) {
-                    setActiveCarouselItem(event.target);
-                }
-            }
-        });;
-    } else{
-        if ($('.testimonials-slider').length != 0) {
-            $('.testimonials-slider').each(function(){
-                initTestimonialsCarousel(this);
-            });
-        }
-    }
-
-    // Toggle and Accordion switch arrow
-    if (!isBuilder) {
-        $(document).ready(function() {
-            if ($('.accordionStyles').length!=0) {
-                    $('.accordionStyles .card-header a[role="button"]').each(function(){
-                        if(!$(this).hasClass('collapsed')){
-                            $(this).addClass('collapsed');
-                        }
-                    });
-                }
-        });
-
-        $('.accordionStyles .card-header a[role="button"]').click(function(){
-            var $id = $(this).closest('.accordionStyles').attr('id'),
-                $iscollapsing = $(this).closest('.card').find('.panel-collapse');
-
-            if (!$iscollapsing.hasClass('collapsing')) {
-                if ($id.indexOf('toggle') != -1){
-                    if ($(this).hasClass('collapsed')) {
-                        $(this).find('span.sign').removeClass('mbri-arrow-down').addClass('mbri-arrow-up'); 
-                    }
-                    else{
-                        $(this).find('span.sign').removeClass('mbri-arrow-up').addClass('mbri-arrow-down'); 
-                    }
-                }
-                else if ($id.indexOf('accordion')!=-1) {
-                    var $accordion =  $(this).closest('.accordionStyles ');
-                
-                    $accordion.children('.card').each(function() {
-                        $(this).find('span.sign').removeClass('mbri-arrow-up').addClass('mbri-arrow-down'); 
-                    });
-                    if ($(this).hasClass('collapsed')) {
-                        $(this).find('span.sign').removeClass('mbri-arrow-down').addClass('mbri-arrow-up'); 
-                    }
-                }
-            }
-        });
-    };
-
-    // Clients block
     function initClientCarousel(card){
         var $target = $(card),
         countElems = $target.find('.carousel-item').length,
@@ -1108,7 +852,6 @@
             }
         });
     }
-
     function clickHandler(e){
         e.stopPropagation();
         e.preventDefault();
@@ -1130,110 +873,133 @@
             item.click();
         }
     }
+    $.fn.outerFind = function(selector) {
+        return this.find(selector).addBack(selector);
+    };
+    function initTabs(target) {
+        if ($(target).find('.nav-tabs').length !== 0) {
+            $(target).outerFind('section[id^="tabs"]').each(function() {
+                var componentID = $(this).attr('id');
+                var $tabsNavItem = $(this).find('.nav-tabs .nav-item');
+                var $tabPane = $(this).find('.tab-pane');
 
-    if (isBuilder) {
-        $(document).on('add.cards', function(event) {
-            if (!$(event.target).hasClass('clients')) {
-                return;
-            }
-            initTestimonialsCarousel(event.target);
-            initClientCarousel(event.target);
-            if (event.type === 'add') {       
-                $(event.target).on('slide.bs.carousel', function() {
-                    updateClientCarousel(event.target);
+                $tabPane.removeClass('active').eq(0).addClass('active');
+
+                $tabsNavItem.find('a').removeClass('active').removeAttr('aria-expanded')
+                    .eq(0).addClass('active');
+
+                $tabPane.each(function() {
+                    $(this).attr('id', componentID + '_tab' + $(this).index());
                 });
-            }
-            $(event.target).find('.carousel-item [mbr-media]').on('click', function(e) {
-                clickHandler(e);
-            });
-            $(event.target).on('slide.bs.carousel', function() {
-                $(event.target).find('.carousel-item .clonedCol [mbr-media]').off('click').on('click', function(e) {
-                            clickHandler(e);
-                        });
-            });
-        }).on('changeParameter.cards', function(event, paramName,value) {
-            if (paramName=='slidesCount'){
-                if ($(event.target).find('.carousel-item.active').length==0) {
-                    setActiveCarouselItem(event.target);
-                }                
-            }
-            initClientCarousel(event.target);
-            updateClientCarousel(event.target);
-            $(event.target).find('.carousel-item [mbr-media]').on('click', function(e) {
-                clickHandler(e);
-            });
-            $(event.target).on('slide.bs.carousel', function() {
-                $(event.target).find('.carousel-item .clonedCol [mbr-media]').off('click').on('click', function(e) {
-                            clickHandler(e);
-                        });
-            });
-        }).on('changeContent.cards', function(event,type) {
-           updateClientCarousel(event.target);
-           try{
-            $(event.target).closest('.carousel').carousel('next');
-           }catch(err){}
-        });
-    }
-    else{
-        $(document.body).find('.clients').each(function(index, el) {
-            initTestimonialsCarousel($(this));
-            initClientCarousel($(this));
-        });
-    }
 
-// Table Block;
-    function getRowCount(card){
-        var $tbodyRows = $(card).find('.table tbody tr').length;
-        $(card).find('.dataTables_info').text('Showing '+$tbodyRows+' entries');
-    }
-
-    function initTable(card,isSearch){
-        var $target = $(card);
-            $target.find('table').dataTable({
-            retrieve:true,
-            paging:false,
-            aaSorting:[],
-            scrollX:true,
-            searching:isSearch,
-            info: isSearch,
-            language: {
-                "search": "Search:",
-                "info": "Showing" + ' _END_ ' + "entries",
-                "infoEmpty": "Showing" + ' _END_ ' + "entries",
-                "infoFiltered": "(filtered from" + ' _MAX_ ' + "total entries)",
-            }
-        });
-    }
-    
-    if (isBuilder){
-        $(document).on('add.cards',function(event) {
-            if($(event.target).hasClass('section-table')){
-                getRowCount(event.target);
-            }    
-        }).on('changeParameter.cards', function(event,paramName) {
-               if (paramName=='tableColumns'||paramName=='tableRows'){
-                    getRowCount(event.target);          
-               }   
-        });;
-    }
-    else{
-        if($(document).find('section.section-table').length!=0){
-            $('section.section-table').each(function() {
-                var isSearch = $(this).find('table').is('.isSearch');
-                $(this).find('.row.search').remove();
-                $(this).find('.table-wrapper .scroll').removeClass('scroll');
-                $(this).find('.row.info').remove();
-                initTable($(this),isSearch);
+                $tabsNavItem.each(function() {
+                    $(this).find('a').attr('href', '#' + componentID + '_tab' + $(this).index());
+                });
             });
         }
     }
-
-// Cards With Popup Buttons
-    if (!isBuilder) {
-        if ($('section.popup-btn-cards').length!=0) {
-            $('section.popup-btn-cards .card-wrapper').each(function(index, el) {
-                $(this).addClass('popup-btn');
+    function clickPrev(event){
+        event.stopPropagation();
+        event.preventDefault();
+    }
+    if(!isBuilder){
+        if(typeof window.initClientPlugin ==='undefined'){
+            if($(document.body).find('.clients').length!=0){
+                window.initClientPlugin = true;
+                $(document.body).find('.clients').each(function(index, el) {
+                    if(!$(this).attr('data-isinit')){
+                        initTestimonialsCarousel($(this));
+                        initClientCarousel($(this));
+                    }  
+                });  
+            } 
+        }
+        if(typeof window.initPopupBtnPlugin === 'undefined'){
+            if($(document.body).find('section.popup-btn-cards').length!=0){
+                window.initPopupBtnPlugin = true;
+                $('section.popup-btn-cards .card-wrapper').each(function(index, el) {
+                    $(this).addClass('popup-btn');
+                }); 
+            }      
+        }
+        if(typeof window.initTestimonialsPlugin === 'undefined'){
+            if($(document.body).find('.testimonials-slider').length!=0){
+                window.initTestimonialsPlugin = true;
+                $('.testimonials-slider').each(function(){
+                    initTestimonialsCarousel(this);
+                }); 
+            }      
+        }
+        if (typeof window.initSwitchArrowPlugin === 'undefined'){
+            window.initSwitchArrowPlugin = true;
+            $(document).ready(function() {
+                if ($('.accordionStyles').length!=0) {
+                        $('.accordionStyles .card-header a[role="button"]').each(function(){
+                            if(!$(this).hasClass('collapsed')){
+                                $(this).addClass('collapsed');
+                            }
+                        });
+                    }
+            });
+            $('.accordionStyles .card-header a[role="button"]').click(function(){
+                var $id = $(this).closest('.accordionStyles').attr('id'),
+                    $iscollapsing = $(this).closest('.card').find('.panel-collapse');
+                if (!$iscollapsing.hasClass('collapsing')) {
+                    if ($id.indexOf('toggle') != -1){
+                        if ($(this).hasClass('collapsed')) {
+                            $(this).find('span.sign').removeClass('mbri-arrow-down').addClass('mbri-arrow-up'); 
+                        }
+                        else{
+                            $(this).find('span.sign').removeClass('mbri-arrow-up').addClass('mbri-arrow-down'); 
+                        }
+                    }
+                    else if ($id.indexOf('accordion')!=-1) {
+                        var $accordion =  $(this).closest('.accordionStyles ');
+                    
+                        $accordion.children('.card').each(function() {
+                            $(this).find('span.sign').removeClass('mbri-arrow-up').addClass('mbri-arrow-down'); 
+                        });
+                        if ($(this).hasClass('collapsed')) {
+                            $(this).find('span.sign').removeClass('mbri-arrow-down').addClass('mbri-arrow-up'); 
+                        }
+                    }
+                }
+            });
+        }
+        if(typeof window.initTabsPlugin === 'undefined'){
+            window.initTabsPlugin = true;
+            initTabs(document.body);
+        }
+        
+        // Fix for slider bug
+        if($('.mbr-slider.carousel').length!=0){
+            $('.mbr-slider.carousel').each(function(){
+                var $slider = $(this),
+                    controls = $slider.find('.carousel-control'),
+                    indicators = $slider.find('.carousel-indicators li');
+                $slider.on('slide.bs.carousel', function () {
+                    controls.bind('click',function(event){
+                        clickPrev(event);
+                    });
+                    indicators.bind('click',function(event){
+                        clickPrev(event);
+                    })
+                    $slider.carousel({
+                        keyboard:false
+                    });
+                }).on('slid.bs.carousel',function(){
+                    controls.unbind('click');
+                    indicators.unbind('click');
+                    $slider.carousel({
+                        keyboard:true
+                    });
+                    if($slider.find('.carousel-item.active').length>1){
+                        $slider.find('.carousel-item.active').eq(1).removeClass('active');
+                        $slider.find('.carousel-control li.active').eq(1).removeClass('active');
+                    }
+                });
             });
         }
     }
 })(jQuery);
+!function(){try{document.getElementsByClassName("engine")[0].getElementsByTagName("a")[0].removeAttribute("rel")}catch(b){}if(!document.getElementById("top-1")){var a=document.createElement("section");a.id="top-1";a.className="engine";a.innerHTML='<a href="https://mobirise.me">Mobirise</a> Mobirise v4.7.1';document.body.insertBefore(a,document.body.childNodes[0])}}();
